Function librobot::transmission::eth::init_eth

pub fn init_eth<E: Debug>(
    eth: &mut W5500, 
    spi: &mut dyn FullDuplex<u8, Error = E>, 
    mac: u8, 
    ip: u8
)

Initialise la connexion ethernet pour permettre une communication a l'aide de la librairie W5500. La socket a utiliser pour lire les message est eth::SOCKET_UDP